requirements.txt
Flask
langchain
langchain-core
langchain-groq
python-dotenv
pydantic